refactor(components): refactor link (#3527)
* refactor(components): refactor link

* fix: buildProp
